9th homework redone
