+++ src/main/java/com/hillel/skoryk/homeworks/lesson9/Apple.java
@@ -1,0 +1,4 @@
+package com.hillel.skoryk.homeworks.lesson9;
+
+class Apple extends Fruit {
+}
