+++ src/main/java/com/hillel/skoryk/homeworks/lesson9/Orange.java
@@ -1,0 +1,4 @@
+package com.hillel.skoryk.homeworks.lesson9;
+
+class Orange extends Fruit {
+}
